requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==4.2.4
django-heroku==0.3.1
gunicorn==21.2.0
packaging==23.1
psycopg2==2.9.7
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.5.0

djangorestframework~=3.14.0
sendgrid~=6.10.0